+++ packages/stk/stk_mesh/stk_mesh/base/FindRestriction.cpp
@@ -52,7 +52,7 @@ const FieldBase::Restriction& find_restriction(const FieldBase& field,
                                                EntityRank erank,
                                                const PartVector& parts)
 {
-  if(field.entity_rank() == erank)
+  if(static_cast<unsigned>(field.entity_rank()) == erank)
   {
       const std::vector<FieldBase::Restriction> & restrictions = field.restrictions();
       for(std::vector<FieldBase::Restriction>::const_iterator it=restrictions.begin(), it_end=restrictions.end(); it != it_end; ++it) {
@@ -69,7 +69,7 @@ const FieldBase::Restriction& find_restriction(const FieldBase& field,
                                                EntityRank erank,
                                                const Part & part)
 {
-  if(field.entity_rank() == erank)
+  if(static_cast<unsigned>(field.entity_rank()) == erank)
   {
       const std::vector<FieldBase::Restriction> & restrictions = field.restrictions();
       for(std::vector<FieldBase::Restriction>::const_iterator it=restrictions.begin(), it_end=restrictions.end(); it != it_end; ++it) {
@@ -90,7 +90,7 @@ const FieldBase::Restriction& find_and_check_restriction(const FieldBase& field,
   const FieldBase::Restriction & empty = empty_field_restriction();
   const FieldBase::Restriction * restriction = & empty;
 
-  if(field.entity_rank() == erank)
+  if(static_cast<unsigned>(field.entity_rank()) == erank)
   {
       const std::vector<FieldBase::Restriction> & restrictions = field.restrictions();
       for(std::vector<FieldBase::Restriction>::const_iterator it=restrictions.begin(), it_end=restrictions.end(); it != it_end; ++it) {
